requirements.txt

pyotp==2.7.0
qrcode==7.3.1
google-generativeai
google-ai-generativelanguage==0.6.10
google-api-core==2.21.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-cloud==0.34.0
google-generativeai
